Extract reducer type definition into separate file
The types were getting _really_ hairy to manage. This stuff is heavily
dependent on closures and each type depends on a ton of other data.

Rather than try to continue carrying types down through the
implementation, I've moved the interface to another file and I'm casting
the implementation to that interface.

Much easier to manage.