Extract `createAction(...)` type signature
Similar to `createReducer(...)` the type signature of
`createAction(...)` was getting in the way and becoming difficult to
manage. I moved the type signature to its own file and added some error
type detection. Speaking of which, have I mentioned how much I love
TypeScript? This sort of stuff would be impossible in Flow.