Implement basic async actions
An async action is an action creator bound to an async function. It
dispatches an optimistic action immediately, then after the function
resolves, you get a more traditional action.

It's far from ready. It still needs type inference, error handling, and
reducer support for optimistic actions. Not to mention working redux
middleware. The current implementation is closer to a proof of concept.