Add half-finished hacker news example
This example is unique. It provides an opportunity to demo async action
creators. It also reveals all the ways retreon breaks when you try to
use async action creators. So... win?

I'm committing this as-is. Later I'll fix some bugs, flesh out more of
the API, and polish the testing experience.