Only override error name given a subclass
If the mixin class is used without another subclass, don't try to set
the error name to `RetreonError` - there's more DX value in the default
`TypeError` or whatever you've passed. It only makes sense to override
the error name if you subclass it again.

This resolves a build failure. For some reason, `this.constructor`
points to the nearest non-anonymous class in the prototypal hierarchy,
but *only* on certain node versions, and *only* in some environments.
I can't repro it locally.

There's probably a sensible explanation, but I simply don't care enough,
and the build failure revealed another (unrelated) problem. Fixing one
should fix the other, so I'll leave the mystery to the gods of chaos.

Amended commit: I changed the approach. Apparently `this.constructor`
was a different reference, but it was inheriting `.name` from the mixin
class. This doesn't perfectly line up, but I suspect the build failures
have something to do with native class support and dual inheritance on
class statics. The `name` property is potentially getting inherited to
anonymous classes.

I swapped to a string comparison instead, with a name much less likely
to conflict in userspace.