Add support for dispatched synchronous iterators
I'll admit, this is a potential footgun. There's never a good reason to
dispatch two actions back to back. The reason I'm adding synchronous
iterators is purely for a post-dispatch hook - I need to be able to run
code once reducers have finished their work without moving to a promise.

This is designed for error handling. If an action throws an error,
I need to trigger the error reducers and then re-throw it. I can't
afford to change the dispatch return type, which is why I'm adding
synchronous iterators.

Let's hope it doesn't come back to bite me.

Resolves #210