Fix regression in example project tests
I used to pass action creators directly to the reducer, but since they
all became iterators, testing is more challenging. For the time being
I've created a real redux store for each test case that uses the actual
middleware. I'd like a nicer solution, but this works well enough for
the short term.