Return values from `yield`
This commit adds return values to `yield` expressions in generators
using the return value of `dispatch`. If you dispatch a plain object,
it's not particularly interesting, but when you consider redux
middleware, the dispatch value becomes much more useful (e.g.
a `GET_STATE` middleware).

Unfortunately, due to a limitation with TypeScript's type inference, I'm
not able to tie the dispatch value back to the `yield` expression type.
I've left the expression as `unknown`, so you'll need to coerce it
manually.