Replace compilation process with tsdx
I've spent quite a bit of time tweaking the compiler settings in this
project, optimizing the distribution files and making sure the bundled
size is reasonable. The results have been underwhelming. While I was
able to reduce the download and bundle size by a small margin, the
compiler toolchain became large, slow, and a pain to maintain. If you're
a JS dev, you know exactly what I'm talking about.

I'm experimenting with the `tsdx` toolchain. It produces optimized
assets for TypeScript libraries which align very closely with what I'm
trying to do. It's relatively popular (which is how I found it), so in
theory, it'll be around for a while.

I like the idea of someone else managing the build process. Like every
JS developer, I'm tired of tweaking compilers. If runtimes come out with
a new optimized distribution standard tomorrow, I'd rather upgrade
`tsdx` than spend the day crying. Let's hope it works in practice.

Note: this commit only implements the build process. Linting and unit
tests are currently broken by the tooling change.