Use tsdx for linting and unit tests
I may regret this down the road, but I've chosen to use the `tsdx`
toolchain for running unit tests and linting the source code. It depends
on some relatively old build tools, so I've hacked newer versions into
it through Yarn Resolutions - as I said, I may regret this later.

The TypeScript parser was old enough that it got tripped up on overload
syntax, and prettier was still on v1. Supposedly those are getting
upgraded with tsdx 0.16.x, and the comments are recent enough that
I trust them on it.

Worst case scenario, I can rip out ESLint and Jest while retaining the
build system.