Upgrade TypeScript
Two main changes: first, if you export a function inline with the
definition and add properties later on in the file, consumers won't see
those in the type signature. Perhaps that's a regression, or perhaps
most people are reasonable and migrate to an `interface` type instead of
inferring after overloads. Second, TS seems to require explicit return
types on iterables or it complains about `any` on `yield` statements.
This only affected tests, since everything else is well described
already.